turns out double werent a good idea

+++ math_utils.c
@@ -8,8 +8,8 @@ float map(double var1, double min1, double max1, double min2, double max2){
 	return min2+(max2-min2)*((var1-min1)/(max1-min1));
 }
 
-double float_rand( double min, double max ){
-	double scale = rand() / (double) RAND_MAX; /* [0, 1.0] */
+float float_rand( float min, float max ){
+	float scale = rand() / (float) RAND_MAX; /* [0, 1.0] */
 	return min + scale * ( max - min );      /* [min, max] */
 }
